refactor : Favorites UI
즐겨찾기 hooks 디렉토리로 일부 변경

#13
